+++ apps/status-page/src/app/(status-page)/[domain]/(private)/layout.tsx
@@ -3,40 +3,17 @@ import {
   FloatingButton,
   StatusPageProvider,
 } from "@/components/status-page/floating-button";
-import { HydrateClient, getQueryClient, trpc } from "@/lib/trpc/server";
 
-export default function Layout({
-  children,
-  params,
-}: {
-  children: React.ReactNode;
-  params: Promise<{ domain: string }>;
-}) {
+export default function Layout({ children }: { children: React.ReactNode }) {
   return (
-    <Hydrate params={params}>
-      <StatusPageProvider>
-        <div className="flex min-h-screen flex-col gap-4">
-          <main className="mx-auto flex w-full max-w-2xl flex-1 flex-col px-3 py-2">
-            {children}
-          </main>
-          <Footer className="w-full border-t" />
-        </div>
-        <FloatingButton />
-      </StatusPageProvider>
-    </Hydrate>
+    <StatusPageProvider>
+      <div className="flex min-h-screen flex-col gap-4">
+        <main className="mx-auto flex w-full max-w-2xl flex-1 flex-col px-3 py-2">
+          {children}
+        </main>
+        <Footer className="w-full border-t" />
+      </div>
+      <FloatingButton />
+    </StatusPageProvider>
   );
-}
-
-async function Hydrate({
-  children,
-  params,
-}: {
-  children: React.ReactNode;
-  params: Promise<{ domain: string }>;
-}) {
-  const queryClient = getQueryClient();
-  await queryClient.prefetchQuery(
-    trpc.statusPage.get.queryOptions({ slug: (await params).domain }),
-  );
-  return <HydrateClient>{children}</HydrateClient>;
 }
